requirements.txt
numpy
selenium
tqdm
pillow
webdriver-manager